# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 经管之星 API
  version: 1.0.0
  description: 前后端唯一接口契约。金额以十进制字符串返回，时间使用 ISO 8601 UTC。
servers:
  - url: /api/v1
security: []
paths:
...
  /questions/frequent:
    get:
      summary: 获取常问问题
      operationId: listFrequentQuestions
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 20, default: 10}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/FrequentQuestion'}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /questions/favorites:
    get:
      summary: 获取收藏问题
      operationId: listFavorites
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/FavoriteQuestion'}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    post:
      summary: 收藏问题
      operationId: addFavorite
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/FavoriteCreate'}}}
      responses:
        '200': {description: Existing, content: {application/json: {schema: {$ref: '#/components/schemas/FavoriteQuestion'}}}}
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/FavoriteQuestion'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /questions/favorites/{favoriteId}:
    parameters:
      - {$ref: '#/components/parameters/FavoriteId'}
    delete:
      summary: 取消收藏问题
      operationId: removeFavorite
      responses:
        '204': {description: Deleted or already absent}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /model-configs:
    get:
      summary: 获取模型配置列表
      operationId: listModels
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/ModelConfig'}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    post:
      summary: 创建模型配置
      operationId: createModel
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ModelConfigCreate'}}}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/ModelConfig'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /model-configs/test:
    post:
      summary: 测试模型连接
      operationId: testModel
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ModelConnectionTestRequest'}}}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/ModelConnectionTestResult'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /model-configs/{id}:
    parameters:
      - {$ref: '#/components/parameters/ModelConfigId'}
    get:
      summary: 获取模型配置详情
      operationId: getModel
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/ModelConfig'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    patch:
      summary: 更新模型配置
      operationId: updateModel
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ModelConfigUpdate'}}}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/ModelConfig'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    delete:
      summary: 删除模型配置
      operationId: deleteModel
      responses:
        '204': {description: Deleted}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /model-configs/{id}/activate:
    parameters:
      - {$ref: '#/components/parameters/ModelConfigId'}
    post:
      summary: 启用模型配置
      operationId: activateModel
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/ModelConfig'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /application-config:
    get:
      summary: 获取应用配置
      operationId: getApplicationConfig
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/ApplicationConfig'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    put:
      summary: 更新应用配置
      operationId: updateApplicationConfig
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ApplicationConfigUpdate'}}}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/ApplicationConfig'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /feedback:
    get:
      summary: 获取反馈列表
      operationId: listFeedback
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: keyword, in: query, schema: {type: string, maxLength: 200}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/FeedbackStatus'}}
        - {name: reason, in: query, schema: {$ref: '#/components/schemas/FeedbackReason'}}
        - {name: userId, in: query, schema: {type: string, maxLength: 64}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/FeedbackListResponse'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    post:
      summary: 提交反馈
      operationId: createFeedback
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/FeedbackCreate'}}}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/FeedbackDetail'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /feedback/{id}:
    parameters:
      - {$ref: '#/components/parameters/FeedbackId'}
    get:
      summary: 获取反馈详情
      operationId: getFeedback
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/FeedbackDetail'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
    patch:
      summary: 处理反馈
      operationId: updateFeedback
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/FeedbackUpdate'}}}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/FeedbackDetail'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /qa/logs:
    get:
      summary: 获取问答日志列表
      operationId: listQaLogs
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: keyword, in: query, schema: {type: string, maxLength: 200}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/ExecutionStatus'}}
        - {name: modelConfigId, in: query, schema: {type: string, format: uuid}}
        - {name: userId, in: query, schema: {type: string, maxLength: 64}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/QaLogListResponse'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
  /qa/logs/{executionId}:
    parameters:
      - {$ref: '#/components/parameters/ExecutionId'}
    get:
      summary: 获取问答日志详情
      operationId: getQaLog
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/QaLogDetail'}}}}
        default: {$ref: '#/components/responses/ErrorResponse'}
...
components:
  parameters:
...
    ExecutionId: {name: executionId, in: path, required: true, schema: {type: string, format: uuid}}
...
    ModelConfigId: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    FeedbackId: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    FavoriteId: {name: favoriteId, in: path, required: true, schema: {type: string, format: uuid}}
...
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
  responses:
    ErrorResponse:
      description: 统一脱敏错误
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
  schemas:
    PageMeta:
      type: object
      required: [page, pageSize, total, hasMore]
      properties:
        page: {type: integer, minimum: 1}
        pageSize: {type: integer, minimum: 1}
        total: {type: integer, minimum: 0}
        hasMore: {type: boolean}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ErrorDetail'}
    ErrorDetail:
      type: object
      required: [code, message, requestId]
      properties:
        code: {type: string, example: VALIDATION_ERROR}
        message: {type: string}
        requestId: {type: string}
        details: {type: object, additionalProperties: true, nullable: true}
...
    ExecutionStatus: {type: string, enum: [queued, running, awaiting_input, completed, failed, cancelled, rejected]}
    IntentType: {type: string, enum: [data_query, clarification, business_definition, product_help, chat, out_of_scope, unsafe]}
...
    StepStatus: {type: string, enum: [pending, running, completed, failed, skipped]}
    ExecutionStepType: {type: string, enum: [intent_classification, clarification_required, schema_selection, sql_generation, sql_validation, query_execution, answer_generation]}
    ExecutionStep:
      type: object
      required: [type, status, summary]
      properties:
        type: {$ref: '#/components/schemas/ExecutionStepType'}
        status: {$ref: '#/components/schemas/StepStatus'}
        summary: {type: string}
        startedAt: {type: string, format: date-time, nullable: true}
        completedAt: {type: string, format: date-time, nullable: true}
        durationMs: {type: integer, minimum: 0, nullable: true}
...
    FrequentQuestion:
      type: object
      required: [question, count, lastAskedAt]
      properties:
        question: {type: string}
        count: {type: integer, minimum: 1}
        lastAskedAt: {type: string, format: date-time}
    FavoriteCreate:
      type: object
      required: [question]
      properties:
        question: {type: string, minLength: 1, maxLength: 2000}
        sourceMessageId: {type: string, format: uuid, nullable: true}
    FavoriteQuestion:
      type: object
      required: [id, question, createdAt]
      properties:
        id: {type: string, format: uuid}
        question: {type: string}
        sourceMessageId: {type: string, format: uuid, nullable: true}
        createdAt: {type: string, format: date-time}
    ModelProvider: {type: string, enum: [openai_compatible]}
    ModelProtocol: {type: string, enum: [responses, chat_completions]}
    ModelConfigCreate:
      type: object
      required: [name, provider, protocol, baseUrl, modelName, apiKey, timeoutSeconds, enabled]
      properties:
        name: {type: string, minLength: 1, maxLength: 50}
        provider: {$ref: '#/components/schemas/ModelProvider'}
        protocol: {$ref: '#/components/schemas/ModelProtocol'}
        baseUrl: {type: string, format: uri, maxLength: 500}
        modelName: {type: string, minLength: 1, maxLength: 100}
        apiKey: {type: string, minLength: 1, writeOnly: true}
        timeoutSeconds: {type: integer, minimum: 1, maximum: 120, default: 30}
        enabled: {type: boolean, default: true}
    ModelConfigUpdate:
      type: object
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 50}
        protocol: {$ref: '#/components/schemas/ModelProtocol'}
        baseUrl: {type: string, format: uri, maxLength: 500}
        modelName: {type: string, minLength: 1, maxLength: 100}
        apiKey: {type: string, minLength: 1, writeOnly: true}
        timeoutSeconds: {type: integer, minimum: 1, maximum: 120}
        enabled: {type: boolean}
    ModelConfig:
      type: object
      required: [id, name, provider, protocol, baseUrl, modelName, apiKeyMask, timeoutSeconds, enabled, isActive, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        provider: {$ref: '#/components/schemas/ModelProvider'}
        protocol: {$ref: '#/components/schemas/ModelProtocol'}
        baseUrl: {type: string, format: uri}
        modelName: {type: string}
        apiKeyMask: {type: string, readOnly: true, example: 'sk-****abcd'}
        timeoutSeconds: {type: integer}
        enabled: {type: boolean}
        isActive: {type: boolean}
        lastTestStatus: {type: string, enum: [success, failed, timeout], nullable: true}
        lastTestedAt: {type: string, format: date-time, nullable: true}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    ModelConnectionTestRequest:
      type: object
      properties:
        modelConfigId: {type: string, format: uuid, nullable: true}
        provider: {$ref: '#/components/schemas/ModelProvider'}
        protocol: {$ref: '#/components/schemas/ModelProtocol'}
        baseUrl: {type: string, format: uri, nullable: true}
        modelName: {type: string, minLength: 1, maxLength: 100, nullable: true}
        apiKey: {type: string, writeOnly: true, nullable: true}
        timeoutSeconds: {type: integer, minimum: 1, maximum: 30, default: 10}
      description: modelConfigId 与完整未保存配置二选一。
    ModelConnectionTestResult:
      type: object
      required: [success, status, message, durationMs]
      properties:
        success: {type: boolean}
        status: {type: string, enum: [success, auth_failed, model_not_found, timeout, unavailable, invalid_config, invalid_response]}
        message: {type: string}
        durationMs: {type: integer, minimum: 0}
    ApplicationConfig:
      type: object
      required: [greetingEnabled, greetingText, recommendedQuestions, followUpEnabled, frequentQuestionsEnabled, frequentQuestionThreshold, modelQaEnabled, ttsEnabled, sttEnabled, version, updatedAt]
      properties:
        greetingEnabled: {type: boolean}
        greetingText: {type: string, maxLength: 1000}
        recommendedQuestions: {type: array, maxItems: 10, items: {type: string, minLength: 1, maxLength: 100}}
        followUpEnabled: {type: boolean}
        frequentQuestionsEnabled: {type: boolean}
        frequentQuestionThreshold: {type: integer, minimum: 1, maximum: 1000}
        modelQaEnabled: {type: boolean}
        ttsEnabled: {type: boolean}
        sttEnabled: {type: boolean}
        version: {type: integer, minimum: 1}
        updatedAt: {type: string, format: date-time}
    ApplicationConfigUpdate:
      type: object
      required: [greetingEnabled, greetingText, recommendedQuestions, followUpEnabled, frequentQuestionsEnabled, frequentQuestionThreshold, modelQaEnabled, ttsEnabled, sttEnabled, version]
      properties:
        greetingEnabled: {type: boolean}
        greetingText: {type: string, maxLength: 1000}
        recommendedQuestions: {type: array, maxItems: 10, items: {type: string, minLength: 1, maxLength: 100}}
        followUpEnabled: {type: boolean}
        frequentQuestionsEnabled: {type: boolean}
        frequentQuestionThreshold: {type: integer, minimum: 1, maximum: 1000}
        modelQaEnabled: {type: boolean}
        ttsEnabled: {type: boolean}
        sttEnabled: {type: boolean}
        version: {type: integer, minimum: 1}
    FeedbackReason: {type: string, enum: [sql_error, result_error, metric_error, answer_error, other]}
    FeedbackStatus: {type: string, enum: [pending, processing, resolved, ignored]}
    FeedbackCreate:
      type: object
      required: [sessionId, assistantMessageId, executionId, reason]
      properties:
        sessionId: {type: string, format: uuid}
        assistantMessageId: {type: string, format: uuid}
        executionId: {type: string, format: uuid}
        reason: {$ref: '#/components/schemas/FeedbackReason'}
        description: {type: string, maxLength: 500, nullable: true}
    FeedbackUpdate:
      type: object
      required: [status, version]
      properties:
        status: {$ref: '#/components/schemas/FeedbackStatus'}
        resolutionNote: {type: string, maxLength: 2000, nullable: true}
        version: {type: integer, minimum: 1}
    FeedbackSummary:
      type: object
      required: [id, userId, question, reason, status, version, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        userId: {type: string, maxLength: 64}
        question: {type: string}
        reason: {$ref: '#/components/schemas/FeedbackReason'}
        description: {type: string, nullable: true}
        status: {$ref: '#/components/schemas/FeedbackStatus'}
        version: {type: integer, minimum: 1}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    FeedbackDetail:
      allOf:
        - {$ref: '#/components/schemas/FeedbackSummary'}
        - type: object
          required: [sessionId, assistantMessageId, executionId, answer]
          properties:
            sessionId: {type: string, format: uuid}
            assistantMessageId: {type: string, format: uuid}
            executionId: {type: string, format: uuid}
            dataSourceNames: {type: array, items: {type: string}}
            sql: {type: string, nullable: true}
            resultSummary: {type: string, nullable: true}
            modelName: {type: string, nullable: true}
            answer: {type: string}
            resolutionNote: {type: string, nullable: true}
    FeedbackListResponse:
      type: object
      required: [items, page]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/FeedbackSummary'}}
        page: {$ref: '#/components/schemas/PageMeta'}
    QaLogSummary:
      type: object
      required: [executionId, userId, requestId, question, status, createdAt]
      properties:
        executionId: {type: string, format: uuid}
        userId: {type: string, maxLength: 64}
        requestId: {type: string}
        question: {type: string}
        status: {$ref: '#/components/schemas/ExecutionStatus'}
        modelName: {type: string, nullable: true}
        rowCount: {type: integer, minimum: 0, nullable: true}
        durationMs: {type: integer, minimum: 0, nullable: true}
        errorCode: {type: string, nullable: true}
        createdAt: {type: string, format: date-time}
    QaLogDetail:
      allOf:
        - {$ref: '#/components/schemas/QaLogSummary'}
        - type: object
          required: [sessionId, userMessageId, intent, normalizedQuestion, missingSlots, clarificationRound, dataSourceNames, selectedObjects, modelCalls, graphVersion, graphNodeTrace, checkpointStatus, ragDocumentIds, ragDegraded, steps]
          properties:
            sessionId: {type: string, format: uuid}
            userMessageId: {type: string, format: uuid}
            assistantMessageId: {type: string, format: uuid, nullable: true}
            intent:
              allOf: [{$ref: '#/components/schemas/IntentType'}]
              nullable: true
            normalizedQuestion: {type: string, nullable: true}
            missingSlots: {type: array, items: {type: string}}
            clarificationRound: {type: integer, minimum: 0, maximum: 2}
            dataSourceNames: {type: array, items: {type: string}}
            selectedObjects: {type: array, items: {type: string}}
            generatedSql: {type: string, nullable: true}
            executedSql: {type: string, nullable: true}
            validationSummary: {type: string, nullable: true}
            tokenUsage: {$ref: '#/components/schemas/TokenUsage'}
            modelCalls: {type: array, items: {$ref: '#/components/schemas/ModelCallAudit'}}
            graphVersion: {type: string, nullable: true}
            graphNodeTrace: {type: array, items: {type: string}}
            checkpointStatus: {type: string, nullable: true}
            ragDocumentIds: {type: array, items: {type: string, format: uuid}}
            ragDegraded: {type: boolean}
            steps: {type: array, items: {$ref: '#/components/schemas/ExecutionStep'}}
            errorMessage: {type: string, nullable: true}
    QaLogListResponse:
      type: object
      required: [items, page]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/QaLogSummary'}}
        page: {$ref: '#/components/schemas/PageMeta'}
    TokenUsage:
      type: object
      required: [promptTokens, completionTokens, totalTokens]
      properties:
        promptTokens: {type: integer, minimum: 0}
        completionTokens: {type: integer, minimum: 0}
        totalTokens: {type: integer, minimum: 0}
    ModelCallAudit:
      type: object
      required: [provider, model, purpose, durationMs, promptTokens, completionTokens, totalTokens, retryCount, status, errorCode]
      properties:
        provider: {type: string}
        model: {type: string}
        purpose: {type: string, enum: [intent_classification, sql_generation, sql_correction, answer_generation]}
        durationMs: {type: integer, minimum: 0}
        promptTokens: {type: integer, minimum: 0}
        completionTokens: {type: integer, minimum: 0}
        totalTokens: {type: integer, minimum: 0}
        retryCount: {type: integer, minimum: 0, maximum: 2}
        status: {type: string, enum: [success, failed, invalid_output]}
        errorCode: {type: string, nullable: true}
...